Перенес dataFilters в store
